Add testing

+++ src/main/kotlin/be/breina/openrgb/OpenRgbClientImpl.kt
@@ -8,11 +8,7 @@ import be.breina.openrgb.model.Device
 import be.breina.openrgb.model.ObjectFactory.device
 import be.breina.openrgb.model.ObjectLinker.link
 import java.awt.Color
-import java.io.BufferedInputStream
-import java.io.BufferedOutputStream
-import java.io.IOException
-import java.io.InputStream
-import java.io.OutputStream
+import java.io.*
 import java.net.Socket
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
@@ -26,10 +22,11 @@ class OpenRgbClientImpl : OpenRgbClient {
     private val outputStream: OutputStream
     private val inputStream: InputStream
     private val protocolVersion: Int
+
     override fun setName(clientName: String) {
         sendMessage(
-                { cb: CommandBuilder -> cb.command(CommandId.SET_CLIENT_NAME) },
-                *(clientName + '\u0000').toByteArray(StandardCharsets.US_ASCII)
+            { cb: CommandBuilder -> cb.command(CommandId.SET_CLIENT_NAME) },
+            *(clientName + '\u0000').toByteArray(StandardCharsets.US_ASCII)
         )
     }
 
@@ -38,6 +35,7 @@ class OpenRgbClientImpl : OpenRgbClient {
             sendMessage({ cb: CommandBuilder -> cb.command(CommandId.REQUEST_PROTOCOL_VERSION) }, MAX_PROTOCOL_VERSION)
             return responseMessage().int
         }
+
     override val controllerCount: Int
         get() {
             sendMessage({ cb: CommandBuilder -> cb.command(CommandId.REQUEST_CONTROLLER_COUNT) })
@@ -46,7 +44,10 @@ class OpenRgbClientImpl : OpenRgbClient {
 
     override fun getControllerData(deviceIndex: Int): Device? {
         assert(deviceIndex >= 0)
-        sendMessage({ cb: CommandBuilder -> cb.command(CommandId.REQUEST_CONTROLLER_DATA).device(deviceIndex) }, protocolVersion)
+        sendMessage(
+            { cb: CommandBuilder -> cb.command(CommandId.REQUEST_CONTROLLER_DATA).device(deviceIndex) },
+            protocolVersion
+        )
         val device = responseDevice()
         link(device, deviceIndex, this)
         return device
@@ -56,27 +57,32 @@ class OpenRgbClientImpl : OpenRgbClient {
         get() = IntStream.range(0, controllerCount).mapToObj(::getControllerData).toArray(::arrayOfNulls)
 
     override fun updateLed(deviceIndex: Int, ledIndex: Int, color: Color) {
-        updateLed(deviceIndex, ledIndex, be.breina.openrgb.model.Color(color.red, color.green, color.blue))
+        updateLed(
+            deviceIndex, ledIndex,
+            be.breina.openrgb.model.Color(color.red.toByte(), color.green.toByte(), color.blue.toByte())
+        )
     }
 
     override fun updateLeds(deviceIndex: Int, vararg colors: Color) {
         updateLeds(
-                deviceIndex,
-                *Arrays.stream(colors).map { color -> be.breina.openrgb.model.Color(color.red, color.green, color.blue) }
-                        .toArray<be.breina.openrgb.model.Color>(::arrayOfNulls)
+            deviceIndex,
+            *Arrays.stream(colors).map { color ->
+                be.breina.openrgb.model.Color(color.red.toByte(), color.green.toByte(), color.blue.toByte())
+            }
+                .toArray<be.breina.openrgb.model.Color>(::arrayOfNulls)
         )
     }
 
     private fun sendMessage(cb: Consumer<CommandBuilder>, data: Int) {
         sendMessage(
-                cb,
-                *ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN).putInt(data).array()
+            cb,
+            *ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN).putInt(data).array()
         )
     }
 
     private fun sendMessage(cb: Consumer<CommandBuilder>, vararg data: Byte) {
         val commandBuilder = CommandBuilder()
-                .size(data.size)
+            .size(data.size)
         cb.accept(commandBuilder)
         val commandHeader = commandBuilder.build()
         try {
@@ -103,27 +109,30 @@ class OpenRgbClientImpl : OpenRgbClient {
     private fun updateLed(deviceIndex: Int, ledIndex: Int, color: be.breina.openrgb.model.Color) {
         assert(deviceIndex >= 0)
         val request = ByteBuffer.allocate(8)
-                .order(ByteOrder.LITTLE_ENDIAN)
-                .putInt(ledIndex)
-                .put(color.red)
-                .put(color.green)
-                .put(color.blue)
+            .order(ByteOrder.LITTLE_ENDIAN)
+            .putInt(ledIndex)
+            .put(color.red)
+            .put(color.green)
+            .put(color.blue)
         request.position(request.position() + 1)
-        sendMessage({ cb: CommandBuilder -> cb.command(CommandId.UPDATE_SINGLE_LED).device(deviceIndex) }, *request.array())
+        sendMessage(
+            { cb: CommandBuilder -> cb.command(CommandId.UPDATE_SINGLE_LED).device(deviceIndex) },
+            *request.array()
+        )
     }
 
     private fun updateLeds(deviceIndex: Int, vararg colors: be.breina.openrgb.model.Color) {
         assert(deviceIndex >= 0)
         val request = ByteBuffer.allocate(4 + 2 + 4 * colors.size)
-                .order(ByteOrder.LITTLE_ENDIAN)
-                .position(4)
-                .putShort(colors.size.toShort())
+            .order(ByteOrder.LITTLE_ENDIAN)
+            .position(4)
+            .putShort(colors.size.toShort())
         Arrays.stream(colors).forEach { color: be.breina.openrgb.model.Color ->
             request
-                    .put(color.red)
-                    .put(color.green)
-                    .put(color.blue)
-                    .position(request.position() + 1)
+                .put(color.red)
+                .put(color.green)
+                .put(color.blue)
+                .position(request.position() + 1)
         }
         sendMessage({ cb: CommandBuilder -> cb.command(CommandId.UPDATE_LEDS).device(deviceIndex) }, *request.array())
     }
